src/main/java/ru/yandex/practicum/filmorate/controller/FriendController.java
package ru.yandex.practicum.filmorate.controller;

import lombok.RequiredArgsConstructor;
import lombok.extern.slf4j.Slf4j;
import org.springframework.web.bind.annotation.*;
import ru.yandex.practicum.filmorate.model.User;
import ru.yandex.practicum.filmorate.service.FriendService;

import java.util.List;

@Slf4j
@RestController
@RequiredArgsConstructor
@RequestMapping("/users")
public class FriendController {
    private final FriendService friendService;

    @PutMapping("/{id}/friends/{friendId}")
    public void addFriend(@PathVariable long id, @PathVariable long friendId) {
        log.info("adding friend id: {} to user id: {}", friendId, id);
        friendService.addFriend(id, friendId);
    }

    @GetMapping("/{id}/friends")
    public List<User> getFriends(@PathVariable long id) {
        log.info("get friends of user id: {}", id);
        return friendService.getFriends(id);
    }

    @DeleteMapping("/{id}/friends/{friendId}")
    public void delFriend(@PathVariable long id, @PathVariable long friendId) {
        log.info("deleting friend id: {} from user id: {}", friendId, id);
        friendService.delFriend(id, friendId);
    }

    @GetMapping("/{id}/friends/common/{otherId}")
    public List<User> getCommonFriends(@PathVariable long id, @PathVariable long otherId) {
        log.info("get common friends of user id {} with other user id {}", id, otherId);
        return friendService.getCommonFriends(id, otherId);
    }
}
